examples/playground/pyproject.toml
# Example app dependencies
# Install with: pip install -e .

[project]
name = "django-agui-playground"
version = "0.1.0"
description = "Django AG-UI Playground Example"
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "django-vite>=3.0",
    "python-dotenv>=1.0",
    "ag-ui-protocol>=0.1.0",
    "pydantic>=2.0",
    "asgiref>=3.7",
    # Framework integrations
    "langgraph>=0.2.0",
    "langchain-openai>=0.1",
    "langchain-core>=0.2",
    "llamaindex>=0.10.0",
    "agno>=2.0.0",
    # LLM providers
    "openai>=1.0.0",
    "azure-openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "foreman>=0.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["example_app"]
